+++ docs/changes.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Add missing schema to Pyramid-URL. This broke ReST rendering on the PyPI page.
 
+- Automatically enable two-phase commit when the SQL dialect supports it.
+
 
 1.0 - July 13, 2014
 -------------------